+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH84M7M7ZHH498GXTGT8Q5VB"
+        "01KH8E1PDH924ANP7915KF7PS8"
     )
     r"""The unique identifier of the data source"""