+++ packages/orq-rc/RELEASES.md
@@ -5569,3 +5569,13 @@ Based on:
 - [python v4.3.0-rc.57] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.57] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.57 - packages/orq-rc
+
+## 2026-02-12 08:06:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.711.0 (2.818.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.58] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.58] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.58 - packages/orq-rc